[WEJBHTTP-139] Refactoring - merge transaction/Utils , naming/Utils & ejb/Utils to common/Utils
