--- transaction/src/main/java/org/wildfly/httpclient/transaction/Utils.java
@@ -1,56 +1,0 @@
-/*
- * JBoss, Home of Professional Open Source.
- * Copyright 2024 Red Hat, Inc., and individual contributors
- * as indicated by the @author tags.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.wildfly.httpclient.transaction;
-
-import org.jboss.marshalling.Marshaller;
-import org.jboss.marshalling.Unmarshaller;
-import org.wildfly.httpclient.common.HttpMarshallerFactory;
-
-import java.util.concurrent.CompletableFuture;
-
-/**
- * @author <a href="mailto:ropalka@redhat.com">Richard Opalka</a>
- */
-final class Utils {
-
-    private Utils() {
-        // forbidden instantiation
-    }
-
-    static <T> Marshaller newMarshaller(final HttpMarshallerFactory factory, final CompletableFuture<T> failureHandler) {
-        Marshaller marshaller = null;
-        try {
-            marshaller = factory.createMarshaller();
-        } catch (Exception e) {
-            failureHandler.completeExceptionally(e);
-        }
-        return marshaller;
-    }
-
-    static <T> Unmarshaller newUnmarshaller(final HttpMarshallerFactory factory, final CompletableFuture<T> failureHandler) {
-        Unmarshaller unmarshaller = null;
-        try {
-            unmarshaller = factory.createUnmarshaller();
-        } catch (Exception e) {
-            failureHandler.completeExceptionally(e);
-        }
-        return unmarshaller;
-    }
-
-}
